Added dynamic support for project cards
