+++ src/components/Projects.js
@@ -1,16 +1,16 @@
 import React from 'react'
 import styled from 'styled-components'
-// import Card from './Card'
-import NewCard from './NewCard'
+import Card from './Card'
 
 /**
  * TODO: fix live source URL for projects array
+ * TODO: move projects data to a separate js file
  */
 
 const projects = [
 	{
 		title: "Tricket",
-		oneLiner: "Tricket is a ticket tracking system.",
+		oneLiner: "Ticket tracking system.",
 		description: `Web application that allows you to create, update, delete and track tickets/tasks for your various projects.
 					 Tricket makes you more productive by giving you a view of your tasks and tickets that needs to be worked on.`,
 		techStack: ["bootstrap", "react", "express", "mongodb"],
@@ -20,10 +20,9 @@ const projects = [
 	},
 	{
 		title: "cornerBet",
-		oneLiner: "cornerBet is an automated bot for scrapping socccer matches and making predictions.",
-		description: `Automated script to scrape soccer matches statistics and make predictions for possible end time corners.
-					cornerBet is built using javascript and powered by cheerio, a library for parsing and manipulating HTML and XML, future updates
-					will include a web version showing prediction statistics and live match tracking.`,
+		oneLiner: "Automated bot for scrapping socccer matches and making predictions.",
+		description: `Automated scrapper for soccer matches statistics and make predictions for possible end time corners.
+					Built using javascript and powered by cheerio, a library for parsing and manipulating HTML and XML`,
 		techStack: ["javascript", "mongodb"],
 		liveSource: "cornerbet",
 		github: "https://github.com/gustavocsak/cornerbet",
@@ -32,15 +31,12 @@ const projects = [
 
 ]
 
-
 const Section = styled.div`
-  	/* height: 100vh; */
-  	//scroll-snap-align: center; 
 	display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: space-between;
-
+	padding: 5rem;
     @media only screen and (max-width: 768px) {
 		height: fit-content;
 		min-height: 100%;
@@ -49,19 +45,15 @@ const Section = styled.div`
 		scroll-snap-align: none;
 	}
 `
+
 const Container = styled.div`
     height: 100%;
-    /* scroll-snap-align: center; // does not align to hero container, liked better this way */
     width: 1400px;
     display: flex;
 	align-items: center;
 	justify-content: center;
-    
-    @media only screen and (max-width: 768px) {
-       /* height: auto; */
+	@media only screen and (max-width: 768px) {
 	   width: 100%;
-	   
-		
     }
 `
 
@@ -84,15 +76,12 @@ const ProjectDisplay = styled.div`
 const Projects = () => {
 	return (
 		<Section>
-
 			<Container>
-
 				<ProjectDisplay>
-					<NewCard />
-					<NewCard />
-					<NewCard />
+					{projects.map((project, index) => {
+						return <Card project={project}/>
+					})}
 				</ProjectDisplay>
-
 			</Container>
 		</Section>
 	)
